+++ src/Former/Form/Form.php
@@ -223,7 +223,13 @@ class Form extends FormerObject
    */
   public function route($name, $params = array())
   {
+    // Set the form action
     $this->action = $this->url->route($name, $params);
+
+    // Set the proper method
+    if ($method = $this->findRouteMethod($name)) {
+      $this->method($method);
+    }
 
     return $this;
   }
@@ -239,6 +245,11 @@ class Form extends FormerObject
   public function controller($name, $params = array())
   {
     $this->action = $this->url->action($name, $params);
+
+    // Set the proper method
+    if ($method = $this->findRouteMethod($name)) {
+      $this->method($method);
+    }
 
     return $this;
   }
@@ -306,6 +317,37 @@ class Form extends FormerObject
   ////////////////////////////////////////////////////////////////////
   ////////////////////////////// HELPERS /////////////////////////////
   ////////////////////////////////////////////////////////////////////
+
+  /**
+   * Find the method of a route by its _uses or name
+   *
+   * @param  string $name
+   *
+   * @return string
+   */
+  protected function findRouteMethod($name)
+  {
+    if (!$this->app->bound('router')) {
+      return;
+    }
+
+    // Get string by name
+    if (!String::contains($name, '@')) {
+      $route = $this->app['router']->getRoutes()->get($name);
+
+    // Get string by uses
+    } else {
+      foreach ($this->app['router']->getRoutes() as $route) {
+        if ($action = $route->getOption('_uses')) {
+          if ($action == $name) {
+            break;
+          }
+        }
+      }
+    }
+
+    return array_get($route->getMethods(), 0);
+  }
 
   /**
    * Apply various parameters according to form type